WIP: Group 8 economy overhaul — population, famine cascade, pricing
Phase 1: Port.Population (100-15K), TargetSupply from population,
famine cascade (starvation → pop decline → prosperity drop),
production scaled by population × prosperity.

Phase 2: Non-linear pricing for essentials (Food/Medicine/Water) —
exponential scarcity curve, wider price bands.

Phase 4: Profit-per-day routing (margin / travelDays, 0.5f intra-region
floor).

Known issue: Deep-time tests failing — ports starving because merchants
aren't delivering food fast enough. Food production rates over-inflated
as band-aid — needs reverting. Correct fix: Phase 3 faction subsidies
to break treasury deadlock so merchants get paid for relief runs.

Mean-reversion band-aid retained with TODO marker.

+++ tests/Ahoy.Simulation.Tests/Infrastructure/TelemetrySnapshot.cs
@@ -11,7 +11,7 @@ public sealed record TelemetrySnapshot
 {
     public int Tick { get; init; }
     public float AvgPortProsperity { get; init; }
-    public int TotalGoldInEconomy { get; init; }
+    public long TotalGoldInEconomy { get; init; }
     public int ActiveShips { get; init; }
     public int ShipsAtSea { get; init; }
     public int AliveIndividuals { get; init; }
@@ -32,9 +32,9 @@ public sealed record TelemetrySnapshot
             Tick = tick,
             AvgPortProsperity = state.Ports.Values.Count > 0
                 ? state.Ports.Values.Average(p => p.Prosperity) : 0,
-            TotalGoldInEconomy = state.Factions.Values.Sum(f => f.TreasuryGold)
-                + state.Individuals.Values.Sum(i => i.CurrentGold)
-                + state.Ships.Values.Sum(s => s.GoldOnBoard),
+            TotalGoldInEconomy = state.Factions.Values.Sum(f => (long)f.TreasuryGold)
+                + state.Individuals.Values.Sum(i => (long)i.CurrentGold)
+                + state.Ships.Values.Sum(s => (long)s.GoldOnBoard),
             ActiveShips = state.Ships.Count,
             ShipsAtSea = state.Ships.Values.Count(s => s.Location is not AtPort),
             AliveIndividuals = state.Individuals.Values.Count(i => i.IsAlive),
